pyproject.toml
[tool.poetry]
name = "lesson27_solution"
version = "0.1.0"
description = ""
authors = ["Albina <aalmukhametova@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
django-admin = "^2.0.1"
psycopg2-binary = "^2.9.5"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
